+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Prospector Changelog
 Version 0.5
 ---
 
+* Files and paths can now be ignored using the `--ignore-paths` and `--ignore-patterns` arguments.
+
 * Full PEP8 compliance can be turned on using the `--full-pep8` flag, which overrides the defaults in the strictness profile.
 * The PEP8 tool will now use existing config if any is found in `.pep8`, `tox.ini`, `setup.cfg` in the path to check, or `~/.config/pep8`. These will override any other configuration specified by Prospector. If none are present, Prospector will fall back on the defaults specified by the strictness.
 * A new flag, `--external-config`, can be used to tweak how PEP8 treats external config. `only`, the default, means that external configuration will be preferred to Prospector configuration. `merge` means that Prospector will combine external configuration and its own